requirements.txt
fastapi
uvicorn
sqlalchemy
jinja2
google-generativeai
pydantic
python-dotenv
requests
beautifulsoup4
